Инструкции любой вложенности (с отладкой)

+++ KismetEditor/Solicen/KismetBytecodeModifier.cs
@@ -1,5 +1,11 @@
-﻿using System;using UAssetAPI;using System.IO;using Newtonsoft.Json.Linq;
-using System.Linq;using System.Collections.Generic;
+﻿using KismetEditor.Solicen;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using UAssetAPI;
 
 namespace Solicen.Kismet
 {
@@ -13,33 +19,58 @@ namespace Solicen.Kismet
             var lines = File.ReadAllLines(filepath);
             foreach(var l in lines)
             {
-                if ("Original | Translation" == l) continue; // Если строка это заголовок
+                if (l.Contains("OriginalText") && l.Contains("Translation")) continue; // Пропускаем строку заголовка
                 if (l.StartsWith("//")) continue; // Если строка начинается с символов комментирования - пропустить
-                var values = l.Split('|');
-                var key = values[0].Trim();
-                var value = values[1].Trim();
 
-                csvValues.Add(key, value);
+                try
+                {
+                    var values = l.Split('|');
+                    var key = values[0].Trim();
+                    var value = values[1].Trim();
+
+                    value = StringHelper.UnEscapeKey(value);
+
+                    try { if (!string.IsNullOrWhiteSpace(key) && !string.IsNullOrWhiteSpace(value)) csvValues.Add(key, value); }
+                    catch (Exception ex) { Console.WriteLine($"[{key}] Элемент с таким ключом уже был добавлен."); }
+                }
+                catch
+                {
+
+                }
+
+
+
+
             }
             return csvValues;
         }
 
         public static void ExtractAndWriteCSV(string assetPath, string _fileName = "")
         {
+
             UAsset asset = new UAsset(assetPath, Version);
             var json = asset.SerializeJson(Newtonsoft.Json.Formatting.Indented);
-            var ubergraph = KismetExtension.GetUbergraphJson(UAsset.DeserializeJson(json));
-            var strings = KismetObject.ToCSV(ubergraph);
+            var ubergraph = KismetExtension.GetUbergraphSerialized(asset);
+
+            if (ProgramProcessor.DebugMode)
+            {
+                var _dummyJson = KismetExtension.GetUbergraphJson(asset); //JsonConvert.SerializeObject(ubergraph, Formatting.Indented);
+                File.WriteAllText($"{Environment.CurrentDirectory}\\{Path.GetFileNameWithoutExtension(assetPath)}_DUMMY.json", _dummyJson.ToString());
+
+            }
+
+            var strings = MapParser.ParseAsCSV(ubergraph);
             _fileName = _fileName == "" ? Path.GetFileNameWithoutExtension(assetPath) : _fileName;
+            if (strings.Length == 0) return;
 
             #region Запись CSV
             var csvFilePath = $"{Environment.CurrentDirectory}\\{_fileName}.csv";
-            
+            Console.WriteLine($"Write CSV to {csvFilePath}");
             if (File.Exists(csvFilePath))
             {
                 // Если CSV уже существует, просто добавить новые строки
                 var csv = File.ReadAllLines(csvFilePath);
-                var lines = strings.Except(csv); lines.ToList().Add("\n");
+                var lines = strings.Except(csv.Select(x => x.Split('|')[0])); lines.ToList().Add("\n");
                 File.AppendAllLines(csvFilePath, lines);
             }           
             else
@@ -51,37 +82,56 @@ namespace Solicen.Kismet
             #endregion
         }
 
+
+        static bool UseBak = true;
         public static void ModifyAsset(string path, Dictionary<string, string> replacement)
         {
+            if (UseBak)
+            {
+                var uexpPath = Path.ChangeExtension(path, ".uexp");
+                if (File.Exists(path+".bak"))
+                {
+                    File.Copy(uexpPath+".bak", uexpPath, true);
+                    File.Copy(path + ".bak", path, true);
+                }
+
+            }
+
             // Загружаем uasset файл с помощью UAssetAPI
             UAsset asset = new UAsset(path, Version);
-            var json = asset.SerializeJson(Newtonsoft.Json.Formatting.Indented);
+            var json = asset.SerializeJson(Formatting.Indented);
             JObject jsonObject = JObject.Parse(json);
+            Console.WriteLine("[Replace string]");
+            var ubergraphExpressions = KismetExtension.GetUbergraphJson(asset);
+            if (ubergraphExpressions == null)
+            {
+                Console.WriteLine("[ERROR] Не удалось получить JArray уберграфа из ассета.");
+                return;
+            }
 
-            var kismet = new KismetExtension(jsonObject);
+            KismetProcessor.ReplaceAll(jsonObject, replacement, asset);
 
-            Console.WriteLine("Replace string");
-            kismet.ReplaceAllInst(replacement);
-            jsonObject = kismet.JsonObject;
+            // --- Отладочный вывод и безопасный режим (без сохранения) ---
 
-            var _json = jsonObject.ToString();
+            var ModdedAsset = UAsset.DeserializeJson(jsonObject.ToString());
+            if (ProgramProcessor.DebugMode)
+            {
+                var _json = jsonObject.ToString();
+                // Сохраняем итоговый JSON в файл для ручной проверки
+                File.WriteAllText($"{Environment.CurrentDirectory}\\Ubergraph.json", _json);
+                Console.WriteLine($"[INFO] Модифицированный JSON сохранен в: {Environment.CurrentDirectory}\\Ubergraph.json");
+            }
+            if (!File.Exists(Path.ChangeExtension(path, ".bak")))
+            {
+                if (File.Exists(path)) File.Copy(path, path + ".bak", true);
+                if (File.Exists(Path.ChangeExtension(path, "uexp"))) File.Copy(Path.ChangeExtension(path, "uexp"), Path.ChangeExtension(path, "uexp") + ".bak", true);
+            }
 
-            if (File.Exists(path)) File.Copy(path, path + ".bak", true);
-            if (File.Exists(Path.ChangeExtension(path, "uexp"))) File.Copy(Path.ChangeExtension(path, "uexp"), Path.ChangeExtension(path, "uexp") + ".bak", true);
-            UAsset.DeserializeJson(_json).Write(path);
-
-            #region Дебаг сегмент для создания Ubergraph текстового файла
-            /*
-            File.WriteAllText($"{Environment.CurrentDirectory}\\JsonModded.json", _json);
-
-            var ubergraph = KismetExtension.GetUbergraphJson(UAsset.DeserializeJson(_json));
-            File.WriteAllText($"{Environment.CurrentDirectory}\\JsonModdedUbergraph.json", ubergraph.ToString());
-            */
-            #endregion
-
-            Console.WriteLine($"Successfully modified [{kismet.ModifiedInst}] instructions.");
+            ModdedAsset.Write(path);
+            Console.WriteLine($"[INFO] Всего было произведено замен: {KismetProcessor.ModifiedInstCount}");
+            Console.WriteLine($"[SUCCESS] Модифицированный ассет успешно сохранен в: {path}");
+            
         }
     }
 
 }
-
